Add suport for python, c, c+=
