+++ inginious/frontend/webapp/templates/layout.html
@@ -62,7 +62,7 @@ $else:
     <script src="$get_homepath()/static/common/js/codemirror/mode/clike/clike.js" type="text/javascript" charset="utf-8"></script>
     <script src="$get_homepath()/static/common/js/codemirror/mode/python/python.js" type="text/javascript" charset="utf-8"></script>
     <script src="$get_homepath()/static/common/js/codemirror/addon/lint/lint.js" type="text/javascript" charset="utf-8"></script>
-    <script src="$get_homepath()/static/common/js/codemirror/addon/lint/java-lint.js" type="text/javascript" charset="utf-8"></script>
+    <script src="$get_homepath()/static/common/js/codemirror/addon/lint/my-linters.js" type="text/javascript" charset="utf-8"></script>
     <script src="$get_homepath()/static/common/js/libs/bootstrap.min.js" type="text/javascript" charset="utf-8"></script>
     <script src="$get_homepath()/static/common/js/libs/bootstrap-datetimepicker.min.js" type="text/javascript" charset="utf-8"></script>
     <script src="$get_homepath()/static/common/js/libs/checked-list-group.js" type="text/javascript" charset="utf-8"></script>
